+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 2.4.12 (21/09/2018)
+
+- [ADD] Seq target (thx @TheSemicolon)
+- [ADD] $ExceptionInfo paramter to Write-Log to add Error tracing
+
 ## 2.4.11 (17 August, 2018)
 
 - Fixed custom targets for locations outside module folder (#20 thx to @jeremymcgee73)